edited1

+++ HotelReservationsManager/Data/Entities/Reservation.cs
@@ -28,7 +28,7 @@ namespace Data.Entities
         public DateTime CheckOutDate { get; set; }
         public bool IncludedBreakfast { get; set; }
         public bool IsAllInclusive { get; set; }
-        public Double FinalPrice { get; set; }
+        public decimal FinalPrice { get; set; }
         public ICollection<ClientReservation> ClientReservations { get; set; }
     }
 }
